Move ibc::query types to relayer
The tx hash query better belongs with chain-specific stuff,
and in any case it's only used in the relayer.
Its presence in ibc modules complicated dependencies.
