+++ relayer/src/link/packet_events.rs
@@ -5,10 +5,10 @@ use tracing::{info, span, trace, warn, Level};
 use ibc::core::ics04_channel::channel::QueryPacketEventDataRequest;
 use ibc::core::ics04_channel::packet::Sequence;
 use ibc::events::{IbcEvent, WithBlockDataType};
-use ibc::query::{QueryBlockRequest, QueryTxRequest};
 use ibc::Height;
 
 use crate::chain::handle::ChainHandle;
+use crate::chain::requests::{QueryBlockRequest, QueryTxRequest};
 use crate::link::error::LinkError;
 use crate::path::PathIdentifiers;
